closes #49
Yaw angle from .arena YAML file is read as double in radians
Yaw angle is CCW
Corrections in CASU scene item orientation

+++ src/qcasusceneitem.h
@@ -48,7 +48,7 @@ public:
      */
     QCasuTreeItem* treeItem;
 
-    QCasuSceneItem(QObject *parent, int x, int y, int yaw, QCasuTreeItem *widget);
+    QCasuSceneItem(QObject *parent, int x, int y, double yaw, QCasuTreeItem *widget);
 
     QRectF boundingRect() const;
